+++ udata/tests/api/test_datasets_api.py
@@ -1329,6 +1329,32 @@ class CommunityResourceAPITest(APITestCase):
                          'new description')
         self.assertTrue(
             CommunityResource.objects.first().url.endswith('test.txt'))
+    
+    def test_community_resource_file_update_old_file_deletion(self):
+        '''It should update a community resource's file and delete the old one'''
+        dataset = VisibleDatasetFactory()
+        user = self.login()
+        community_resource = CommunityResourceFactory(dataset=dataset,
+                                                      owner=user)
+        response = self.post(
+            url_for('api.upload_community_resource',
+                    community=community_resource),
+            {'file': (BytesIO(b'aaa'), 'test.txt')}, json=False)
+        self.assert200(response)
+        data = json.loads(response.data)
+        self.assertEqual(data['id'], str(community_resource.id))
+        self.assertEqual(data['title'], 'test.txt')
+
+        response = self.post(
+            url_for('api.upload_community_resource',
+                    community=community_resource),
+            {'file': (BytesIO(b'aaa'), 'test_update.txt')}, json=False)
+        self.assert200(response)
+        data = json.loads(response.data)
+        self.assertEqual(data['id'], str(community_resource.id))
+        self.assertEqual(data['title'], 'test-update.txt')
+
+        self.assertEqual(len(list(storages.resources.list_files())), 1)
 
     def test_community_resource_api_create_remote(self):
         '''It should create a remote community resource from the API'''